upload images and adding pagination

+++ controllers/admin.js
@@ -1,9 +1,8 @@
 const Product = require("../models/product");
-const mongodb = require("mongodb");
 const { validationResult } = require("express-validator");
-const { default: mongoose } = require("mongoose");
+const fileHelper = require("../util/file");
 
-const ObjectId = mongodb.ObjectId.createFromHexString;
+const ITEMS_PER_PAGE = 3;
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -18,10 +17,24 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
 
+  // Multer filter the file and is not an image => View error message
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add product",
+      path: "/admin/add-product",
+      editing: false,
+      hasError: true,
+      product: { title, price, description },
+      errorMessage: "Attached file is not an image",
+      validationErrors: [],
+    });
+  }
+
+  // Input Validations
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
@@ -29,11 +42,13 @@ exports.postAddProduct = (req, res, next) => {
       path: "/admin/add-product",
       editing: false,
       hasError: true,
-      product: { title, imageUrl, price, description },
+      product: { title, price, description },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array(),
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     title,
@@ -42,25 +57,12 @@ exports.postAddProduct = (req, res, next) => {
     description,
     userId: req.user,
   });
-  // throw new Error('Dummy')
   product
     .save()
     .then((result) => {
       res.redirect("/admin/products");
     })
     .catch((err) => {
-      // return res.status(500).render("admin/edit-product", {
-      //   pageTitle: "Add product",
-      //   path: "/admin/add-product",
-      //   editing: false,
-      //   hasError: true,
-      //   product: { title, imageUrl, price, description },
-      //   errorMessage: "Database operation failed. Please try again",
-      //   validationErrors: [],
-      // });
-
-      // return res.redirect('/500')
-
       const error = new Error(err);
       error.httpStatusCode = 500;
       next(error);
@@ -96,7 +98,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const productId = req.body.productId;
   const updatedTitle = req.body.title;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedPrice = req.body.price;
   const updatedDescription = req.body.description;
 
@@ -108,7 +110,6 @@ exports.postEditProduct = (req, res, next) => {
       editing: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDescription,
         _id: productId,
@@ -124,7 +125,10 @@ exports.postEditProduct = (req, res, next) => {
       return res.redirect("/");
     }
     product.title = updatedTitle;
-    product.imageUrl = updatedImageUrl;
+    if (image) {
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = image.path;
+    }
     product.price = updatedPrice;
     product.description = updatedDescription;
     return product
@@ -142,8 +146,15 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId;
-  Product.deleteOne({ _id: productId, userId: req.user._id })
-    .then((resutl) => {
+  Product.findById(productId)
+    .then((product) => {
+      if (!product) {
+        throw new Error("Product not found.");
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: productId, userId: req.user._id });
+    })
+    .then(() => {
       res.redirect("/admin/products");
     })
     .catch((err) => {
@@ -154,14 +165,27 @@ exports.postDeleteProduct = (req, res, next) => {
 };
 
 exports.getAdminProducts = (req, res, next) => {
-  Product.find({ userId: req.user._id })
-    // .select('title price -_id')
-    // .populate('userId', 'name')
+  const page = +req.query.page || 1;
+  let totalItems;
+
+  Product.countDocuments()
+    .then((itemsCount) => {
+      totalItems = itemsCount;
+      return (
+        Product.find({ userId: req.user._id })
+          // .select('title price -_id')
+          // .populate('userId', 'name')
+          .skip((page - 1) * ITEMS_PER_PAGE)
+          .limit(ITEMS_PER_PAGE)
+      );
+    })
     .then((products) => {
       res.render("admin/products", {
         products: products,
         pageTitle: "Admin Products",
         path: "/admin/products",
+        pageNums: Math.ceil(totalItems / ITEMS_PER_PAGE),
+        currentPage: page,
       });
     })
     .catch((err) => {
